tasks: optionally turn skips into failures in cephfs tests
Since if you're running our carefully crafted suite
YAMLs, nothing should be skipped and we'd like to know
about it if that isn't the case.  To avoid cluttering
the YAML files with the option, default to raising
these as failures: use the fail_on_skip:false option
when running these by hand.

Signed-off-by: John Spray <john.spray@redhat.com>